requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
PyJWT==2.3.0
pytz==2022.1
qrcode==7.3.1
sqlparse==0.4.2
tzdata==2022.1
